split physics functionality into its own script
